requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
alembic==1.13.1

# Redis
redis[hiredis]==5.0.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration
pydantic-settings==2.0.3
python-dotenv==1.0.0
email-validator==2.1.0

# HTTP client
httpx==0.25.2

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code quality
black==23.11.0
isort==5.12.0
flake8==6.1.0

# External integrations (will be used in later tasks)
stripe==7.8.0
openai==1.3.7